LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/url-encoded-expected.txt



PASS text.simple 
PASS textarea.simple 
FAIL nokeygen.simple assert_equals: expected "foo=barb" but got "foo=barb&=MIICQTCCASkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDUQG7OUET%2BonWyar%2BS%2Bu6sZraJSp7Z165cAcpCXJfswpnSETvI8TnamSmO4Y5jZkavWxDxhoztrM4Q5c72Uufqx4aay9wmWEYlj7ytorSuCkG8tbKQEGWLh%2Fm2M5l3B4Lvz4PTka73%2BzJxm%2FeDyXAKkp1efF9QTOdsAMMQJpoLikxx%2BlFfLGgbqKp5RRYneOUx9d4lpDhW8PnaWdm19TL3RgZaKhvJOcAtrCoeW%2B%2F54zGhjhWeHVCZC%2B%2FsHsrjnDcuNiCC%2FtkCi9PzhZoPVikx4RLS4ws9hWTRnD%2BSaAXAGTbiPBbyTHWTnG6%2Ff8gTu88j0Pr34gfQExjHKUDfka7tAgMBAAEWAQAwDQYJKoZIhvcNAQEEBQADggEBAM6%2FQV3%2BvdaBxTuvl3QR6OQHHRGtCmnfnRzENK%2FjIUUBSJVJJl3BqwiD7G7few7LoCIdGR3EpZUB6QmOQzMcMBi7ymOAyyx7iw%2F2iZjUooxUOLTvAspHRmsTYAG9VtHYzVZgRKOhypFftcdqRhWliXNW2X254eVW2YWpP5hT%2Fartbqjd7Tv7jN5lLuKmTxTI35JiHw7L8jRdIBgP5G0bu2EG8Fvk7%2FbqVlhdlyHg7XJD6CsB0pU%2FXj3mor7DtTLZ9rwifu3b7Ib%2F03Ak9OArav4KxMo7gb7gagFgjStt9WXVSUsa%2BedrfezVccWCT6ZFDVEJii0zBl2d9FQKRunonPk%3D"
